backend/pyproject.toml
[tool.poetry]
name = "ponder-backend"
version = "1.0.0"
description = "AI-powered learning platform backend"
authors = ["Ponder Team <team@ponder.ai>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.109.0"
uvicorn = {extras = ["standard"], version = "0.27.0"}
python-multipart = "0.0.6"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
passlib = {extras = ["bcrypt"], version = "1.7.4"}
python-dotenv = "1.0.0"
openai = "1.3.0"
sqlalchemy = "2.0.39"
pydantic-settings = "2.7.1"
pydantic = "^2.0.0"
numpy = "^1.26.0"
pandas = "^1.5.0"
requests = "2.31.0"
psutil = "5.9.8"
tenacity = "8.2.3"
redis = "^5.0.1"
aioredis = "^2.0.0"
fastapi-limiter = "^0.1.5"
websockets = "12.0"
slowapi = "0.1.9"
psycopg2-binary = "2.9.10"
prometheus-client = "0.16.0"
aiofiles = "23.1.0"
qdrant-client = "1.7.3"
alembic = "^1.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "7.4.3"
pytest-asyncio = "0.21.1"
pytest-mock = "3.12.0"
httpx = "0.25.2"
pytest-cov = "4.1.0"
factory-boy = "3.3.0"
black = "^23.0.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
mypy = "^1.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "qdrant_client.*",
    "slowapi.*",
    "fastapi_limiter.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"